feat: decode encode loop architecture complete
only file that should have errors right now is translations, which will probably be extremely similar to the current cli layout. need to work on getting that structure figured out. btw there's still a borrow error on the next key assignment in the b+ tree splitting and merging logic

+++ server/src/cli.rs
@@ -32,7 +32,7 @@ impl Call {
             ["SET", key, val, type_tag] => {
                 info!("Input parsed as typed SET");
                 let value: Option::<Value> = match *type_tag {
-                    "int" => Some(Value::Int(val.parse::<i64>().map_err(|_| DbError::BadVal)?)),
+                    "int" => Some(Value::Int(val.parse::<isize>().map_err(|_| DbError::BadVal)?)),
                     "float" => Some(Value::Float(val.parse::<f64>().map_err(|_| DbError::BadVal)?)),
                     "text" => Some(Value::Text(val.to_string())),
                     "blob" => {
